backend/requirements.txt
fastapi[standard]
uvicorn[standard]
google-genai
python-dotenv
httpx
